修复SwitchVehicleWeaponMessage编码错误

+++ src/main/java/com/atsuishio/superbwarfare/network/message/SwitchVehicleWeaponMessage.java
@@ -19,7 +19,7 @@ public class SwitchVehicleWeaponMessage {
     }
 
     public static void encode(SwitchVehicleWeaponMessage message, FriendlyByteBuf byteBuf) {
-        byteBuf.writeDouble(message.index);
+        byteBuf.writeInt(message.index);
         byteBuf.writeDouble(message.scroll);
     }
